+++ docs-beta/src/components/NavBar.tsx
@@ -54,7 +54,7 @@ export default function NavBar() {
             <img className="w-20 h-7" src="img/lage.png" alt="Logo" />
           </a>
           <NavButton to="/docs/Introducing Lage/Overview">Guide</NavButton>
-          <NavButton to="/lage/flow-index">Grapher</NavButton>
+          <NavButton to="/lage/grapher">Grapher</NavButton>
         </div>
         {/*Aligned to right*/}
         <div className="flex items-center">
