+++ docs-beta/src/components/flow-graph/IndependentNode.tsx
